pyproject.toml
[project]
name = "multi-source-rag"
version = "1.0.0"
description = "Multi-Source RAG Pipeline API with PDF, CSV, and Web content support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.18",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-groq>=0.2.0",
    "pypdf>=5.0.0",
    "pymupdf>=1.24.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "html5lib>=1.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "werkzeug>=3.1.5",
]
